fix: renamed appUsersCubit to profilesCubit
