--- lib/cubits/app_user/app_user_state.dart
@@ -1,22 +1,0 @@
-part of 'app_user_cubit.dart';
-
-@immutable
-abstract class AppUserState {}
-
-class AppUserInitial extends AppUserState {}
-
-class AppUserLoaded extends AppUserState {
-  AppUserLoaded({
-    required this.appUsers,
-    required this.self,
-  });
-
-  final Map<String, Profile?> appUsers;
-  final Profile self;
-}
-
-/// State where the logged in user has not been registered yet.
-class AppUserNoProfile extends AppUserState {}
-
-/// State where the user is updating the profile
-class AppUserUpdating extends AppUserState {}
