+++ lib/main.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:supabase_chat/cubits/profiles/profiles_cubit.dart';
 import 'package:supabase_chat/utils/constants.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
-import 'package:supabase_chat/cubits/app_user/app_user_cubit.dart';
 import 'package:supabase_chat/pages/splash_page.dart';
 
 Future<void> main() async {
@@ -22,8 +22,8 @@ class MyApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return BlocProvider<AppUserCubit>(
-      create: (context) => AppUserCubit(),
+    return BlocProvider<ProfilesCubit>(
+      create: (context) => ProfilesCubit(),
       child: MaterialApp(
         title: 'SupaChat',
         theme: appTheme,
